Anaerobic digestion study

This study explored the microbial indicators that could improve the efficacy of anaerobic digestion (AD) bioprocess and prevent its failure. The samples were treated with two different ranges of phenol concentration (effect of interest) and processed at five different dates (batch effect). This study includes a clear and strong batch effect with an approx. balanced batch x treatment design.

data('AD_data')

Format

A list containing three sub-lists of data FullData, EgData and CorrectData:

FullData

A list containing three data sets: X.count which are the raw data represented as a matrix with 75 samples and 567 OTUs; metadata which is a data frame containing the meta data information of samples in X.count; taxa which is a data frame containing the taxonomy of each OTU in X.count.

EgData

A list containing three data sets: X.clr which are the filtered and centered log ratio transformed data including 75 samples and 231 OTUs from the raw data in the FullData list; Y.trt which is a factor of phenol concentrations for each sample that is the effect of interest in the AD study; Y.bat which is a factor of sample processing dates for each sample treated as the batch effect.

CorrectData

A list containing seven data sets before or after batch effect correction using different methods. Each data set includes 75 samples and 231 OTUs.

Source

The raw data were provided by Dr. Olivier Chapleur and published at the referenced article. Filtering and normalisation described in our package vignette.

Value

None.

References

Chapleur O, Madigou C, Civade R, Rodolphe Y, Mazéas L, Bouchez T (2016). “Increasing concentrations of phenol progressively affect anaerobic digestion of cellulose and associated microbial communities.” Biodegradation, 27(1), 15--27.