Partial Least Squares Discriminant Analysis for Batch Effect Correction

This function removes batch variation from the input data given batch grouping information and the number of associated components.

PLSDA_batch(
  X,
  Y.trt = NULL,
  Y.bat,
  ncomp.trt = 2,
  ncomp.bat = 2,
  keepX.trt = rep(ncol(X), ncomp.trt),
  keepX.bat = rep(ncol(X), ncomp.bat),
  max.iter = 500,
  tol = 1e-06,
  near.zero.var = TRUE,
  balance = TRUE
)

Arguments

X
A numeric matrix as an explanatory matrix. NAs are not allowed.

Y.trt
A factor or a class vector for the treatment grouping information (categorical outcome variable). Without the input of Y.trt, treatment variation cannot be preserved before correcting for batch effects.

Y.bat
A factor or a class vector for the batch grouping information (categorical outcome variable).

ncomp.trt
Integer, the number of treatment associated dimensions to include in the model.

ncomp.bat
Integer, the number of batch associated dimensions to include in the model.

keepX.trt
A numeric vector of length ncomp.trt, the number of variables to keep in \(X\)-loadings. By default all variables are kept in the model. A valid input of keepX.trt extends PLSDA_batch to a sparse version.

keepX.bat
A numeric vector of length ncomp.bat, the number of variables to keep in \(X\)-loadings. By default all variables are kept in the model. We usually use default setting.

max.iter
Integer, the maximum number of iterations.

tol
Numeric, convergence stopping value.

near.zero.var
Logical, should be set to TRUE in particular for data with many zero values. Setting this argument to FALSE (when appropriate) will speed up the computations. Default value is TRUE.

balance
Logical, should be set to TRUE, if the batch x treatment design is balanced (or complete). Setting this argument to FALSE extends PLSDA_batch to weighted PLSDA_batch. Default value is TRUE.

Value

PLSDA_batch returns a list that contains the following components:

X

The original explanatory matrix X.

X.nobatch

The batch corrected matrix with the same dimension as the input matrix.

X.notrt

The matrix from which treatment variation is removed.

Y

The original outcome variables Y.trt and Y.bat.

latent_var.trt

The treatment associated latent components calculated with corresponding latent dimensions.

latent_var.bat

The batch associated latent components calculated with corresponding latent dimensions.

loadings.trt

The estimated treatment associated latent dimensions.

loadings.bat

The estimated batch associated latent dimensions.

tol

The tolerance used in the iterative algorithm, convergence stopping value.

max.iter

The maximum number of iterations.

iter.trt

Number of iterations of the algorthm for each treatment associated component.

iter.bat

Number of iterations of the algorthm for each batch associated component.

explained_variance.trt

The amount of data variance explained per treatment associated component.

explained_variance.bat

The amount of data variance explained per batch associated component.

weight

The sample weights, all \(1\) for a balanced batch x treatment design.

References

Wang Y, Lê Cao K (2020). “A multivariate method to correct for batch effects in microbiome data.” bioRxiv.

See also

Author

Yiwen Wang, Kim-Anh Lê Cao

Examples

## First example
## PLSDA-batch
data('AD_data')
X = AD_data$EgData$X.clr # centered log ratio transformed data
Y.trt = AD_data$EgData$Y.trt # treatment information
Y.bat = AD_data$EgData$Y.bat # batch information
ad_plsda_batch <- PLSDA_batch(X, Y.trt, Y.bat, ncomp.trt = 1, ncomp.bat = 5)
ad_X.corrected <- ad_plsda_batch$X.nobatch # batch corrected data

## Second example
## sparse PLSDA-batch
ad_splsda_batch <- PLSDA_batch(X, Y.trt, Y.bat, ncomp.trt = 1,
                               keepX.trt = 30, ncomp.bat = 5)

## Third example
## weighted PLSDA-batch
## The simulated data with an unbalanced batch x treatment design
i = 1
simulation <- simData(bat.mean = 7, sd.bat = 8, trt.mean = 3, sd.trt = 2,
                       N = 40, p_total = 300, p_trt_relevant = 60,
                       p_bat_relevant = 150, percentage_samples = 0.2,
                       percentage_overlap_variables = 0.5)

sdata_wplsda_batch <- PLSDA_batch(X = simulation$data,
                                  Y.trt = simulation$Y.trt,
                                  Y.bat = simulation$Y.bat,
                                  ncomp.trt = 1,
                                  ncomp.bat = 1,
                                  balance = FALSE)

## Fourth example
## sparse weighted PLSDA-batch
sdata_swplsda_batch <- PLSDA_batch(X = simulation$data,
                                   Y.trt = simulation$Y.trt,
                                   Y.bat = simulation$Y.bat,
                                   ncomp.trt = 1,
                                   keepX.trt = length(simulation$true.trt),
                                   ncomp.bat = 1,
                                   balance = FALSE)