Principal Component Analysis (PCA) with Density Plots per Component

This function draws a PCA sample plot with density plots per principal component.

Scatter_Density(
  object,
  batch = NULL,
  trt = NULL,
  xlim = NULL,
  ylim = NULL,
  batch.legend.title = "Batch",
  trt.legend.title = "Treatment",
  density.lwd = 0.2,
  title = NULL,
  title.cex = 1.5,
  legend.cex = 0.7,
  legend.title.cex = 0.75
)

Arguments

object
The object of class PCA.

batch
A factor or a class vector for the batch grouping information (categorical outcome variable).

trt
A factor or a class vector for the treatment grouping information (categorical outcome variable).

xlim
A numeric vector of length 2, indicating the x coordinate ranges.

ylim
A numeric vector of length 2, indicating the y coordinate ranges.

batch.legend.title
Character, the legend title of batches.

trt.legend.title
Character, the legend title of treatments.

density.lwd
Numeric, the thickness of density lines.

title
Character, the plot title.

title.cex
Numeric, the size of plot title.

legend.cex
Numeric, the size of legends.

legend.title.cex
Numeric, the size of legend title.

Value

none

Examples

data('AD_data')
ad.clr <- AD_data$EgData$X.clr
ad.pca.before <- pca(ad.clr, ncomp = 3, scale = TRUE)

#> Error in pca(ad.clr, ncomp = 3, scale = TRUE): could not find function "pca"

ad.batch = AD_data$EgData$Y.bat
ad.trt = AD_data$EgData$Y.trt
Scatter_Density(project = ad.pca.before, batch = ad.batch, trt = ad.trt)

#> Error in Scatter_Density(project = ad.pca.before, batch = ad.batch, trt = ad.trt): unused argument (project = ad.pca.before)