Alignment Scores for Evaluating the Degree of Mixing Samples

This function evaluates the degree of mixing samples from different batches in the batch corrected data. It is based on the dissimilarity matrix from Principal Component Analysis.

alignment_score(
  data,
  batch,
  var = 0.95,
  k = round(0.1 * nrow(data)),
  ncomp = 20
)

Arguments

data

A numeric matrix. Samples are in rows, while variables are in columns. NAs are not allowed.

batch

A factor or a class vector for the batch grouping information (categorical outcome variable). The length should be equal to the number of samples in the data.

var

The proportion of data variance explained by the principal components, ranging from 0 to 1. Default value is 0.95.

k

Integer, the number of nearest neighbours. By default 10% of the number of samples are used.

ncomp

Integer, the number of components for principal component analysis. Default value is 20.

Value

A numeric alignment score that ranges from 0 to 1, representing poor to perfect performance of mixing the samples from different batches.

References

Butler A, Hoffman P, Smibert P, Papalexi E, Satija R (2018). “Integrating single-cell transcriptomic data across different conditions, technologies, and species.” Nature biotechnology, 36(5), 411--420.

See also

Author

Yiwen Wang, Kim-Anh Lê Cao

Examples

data('sponge_data')
X = sponge_data$X.clr # centered log ratio transformed data
batch = sponge_data$Y.bat # batch information

alignment_score(data = X, batch = batch, var = 0.95, k = 3, ncomp = 20)
#> [1] 0.05376344