Matrix Deflation

This function removes the variance of given component t from the input matrix X. $$\hat{X} = X - t (t^{\top}t)^{-1}t^{\top}X$$ It is a built-in funciton of PLSDA_batch.

deflate_mtx(X, t)

Arguments

X
A numeric matrix to be deflated. It assumes that samples are on the row, while variables are on the column. NAs are not allowed.

t
A component to be deflated out from the matrix.

Value

A deflated matrix with the same dimension as the input matrix.

References

Barker M, Rayens W (2003). “Partial least squares for discrimination.” Journal of Chemometrics: A Journal of the Chemometrics Society, 17(3), 166--173.

Examples

set.seed(123)
library(mixOmics) # pca
X <- matrix(rnorm(100), ncol = 20)
## pca for original matrix
pca.X <- pca(X)
## deflate by the first PC and perform pca on the deflated matrix
t <- pca.X$variates$X[,1]
X.deflate <- deflate_mtx(X, t)
pca.X.deflate <- mixOmics::pca(X.deflate)
## see if the first PC of he deflated matrix
## equals the second original pc in terms of absolute value
pca.X.deflate$variates$X[,1]

#>          1          2          3          4          5 
#> -2.0265641 -0.7937048 -1.7010822  2.4479385  2.0734125 

# -2.0265641 -0.7937048 -1.7010822  2.4479385  2.0734125
pca.X$variates$X[,2]

#>          1          2          3          4          5 
#>  2.0265641  0.7937048  1.7010822 -2.4479385 -2.0734125 

#  2.0265641  0.7937048  1.7010822 -2.4479385 -2.0734125