Density Plot

This function draws an overlap of multiple density plots for each batch.

density_plot(
  df,
  title = NULL,
  batch.legend.title = "Batch",
  xlab = "Value",
  title.hjust = 0.5
)

Arguments

df
A data frame used to draw the density plots.

title
Character, the plot title.

batch.legend.title
Character, the legend title of batches.

xlab
Character, x-axis title.

title.hjust
Numeric, horizontal justification of the plot title, in the range of \(0\) to \(1\).

Value

none

Examples

data('AD_data')
ad.clr <- AD_data$EgData$X.clr
ad.batch = AD_data$EgData$Y.bat
ad.df <- data.frame(value = ad.clr[,1], batch = ad.batch)
density_plot(df = ad.df, title = 'OTU 12')

#> Error in aes(x = value, fill = batch): could not find function "aes"