Percentile Normalisation

This function corrects for batch effects in case-control microbiome studies. Briefly, the relative abundance of microbial variables (i.e. bacterial taxa) in case (i.e. disease) samples are converted to percentiles of the equivalent variables in control (i.e. healthy) samples within a batch prior to pooling data across batches. Pooled batches must have similar case and control cohort definitions.

Usage,

percentile_norm(data = data, batch = batch, trt = trt, ctrl.grp)

Arguments

data

A data frame that contains the microbial variables and required to be corrected for batch effects. Samples as rows and variables as columns.

batch

A factor or a class vector for the batch grouping information (categorical outcome variable).

trt

A factor or a class vector for the treatment grouping information (categorical outcome variable).

ctrl.grp

Character, the name of control samples (i.e. healthy).

Value

A data frame that corrected for batch effects.

References

Gibbons SM, Duvallet C, Alm EJ (2018). “Correcting for batch effects in case-control microbiome studies.” PLoS Computational Biology, 14(4), e1006102.

See also

Author

Yiwen Wang, Kim-Anh Lê Cao

Examples

data('AD_data')
ad.clr <- AD_data$EgData$X.clr # centered log ratio transformed data
ad.batch = AD_data$EgData$Y.bat # batch information
ad.trt = AD_data$EgData$Y.trt # treatment information
ad.PN <- percentile_norm(data = ad.clr, batch = ad.batch,
                         trt = ad.trt, ctrl.grp = '0-0.5')