Data Simulation Based on Components and Multivariate Strategy

This function simulates data with treatment or/and batch effects under different batch x treatment design using component-based and multivariate strategy.

Usage,

simData(
  bat.mean,
  sd.bat = 1,
  trt.mean,
  sd.trt = 1,
  noise = 0.2,
  N,
  p_total,
  p_trt_relevant,
  p_bat_relevant,
  percentage_samples = 0.5,
  percentage_overlap_variables = 0.5,
  seeds = round(runif(1, 0, 999))
)

Arguments

bat.mean

Numeric, the absolute mean of two batches respectively in batch associated components.

sd.bat

Numeric, the variability of batch effects.

trt.mean

Numeric, the absolute mean of two treatment groups respectively in treatment associated components.

sd.trt

Numeric, the variability of treatment effects.

noise

Numeric, the variability (standard deviation) of background noise.

N

Integer, the number of samples.

p_total

Integer, the number of variables.

p_trt_relevant

Integer, the number of variables affected with treatment effects.

p_bat_relevant

Integer, the number of variables affected with batch effects.

percentage_samples

Numeric, in the range of \(0\) to \(1\), the percentage of samples in the first batch and first treatment group.

percentage_overlap_variables

Numeric, in the range of \(0\) to \(1\), the percentage of variables with both batch and treatment effects given fixed number of variables with each effect.

seeds

Integer. Default value is randomly selected from \(0\) to \(999\).

Value

simData returns a list that contains the following components:

data

The simulated data.

cleanData

The ground-truth data without the batch effect.

Y.trt

The simulated outcome variable indicating the treatment grouping information.

Y.bat

The simulated outcome variable indicating the batch grouping information.

true.trt

The variables assigned with the treatment effect.

true.batch

The variables assigned with the batch effect.

tComp_relevant

The component simulated with the treatment effect.

bComp_relevant

The component simulated with the batch effect.

References

Wang Y, Lê Cao K (2020). “A multivariate method to correct for batch effects in microbiome data.” bioRxiv.

Author

Yiwen Wang, Kim-Anh Lê Cao

Examples

## First example
## Balanced batch x treatment design
i = 1
simulation1 <- simData(bat.mean = 7, sd.bat = 8, trt.mean = 3, sd.trt = 2,
                       N = 40, p_total = 300, p_trt_relevant = 60,
                       p_bat_relevant = 150,
                       percentage_samples = 0.5,
                       percentage_overlap_variables = 0.5)

## Second example
## Unbalanced batch x treatment design
i = 1
simulation2 <- simData(bat.mean = 7, sd.bat = 8, trt.mean = 3, sd.trt = 2,
                       N = 40, p_total = 300, p_trt_relevant = 60,
                       p_bat_relevant = 150,
                       percentage_samples = 0.2,
                       percentage_overlap_variables = 0.5)