Data Simulation from Multivariate Negative Binomial Distribution

This function simulates data with treatment and batch effects under different batch x treatment designs from multivariate negative binomial distribution using quantile-quantile transformation between multivariate normal and negative binomial distribution. This function can generate two or three batch groups.

simData_mnegbinom(
  batch.group = 2,
  mean.batch,
  sd.batch = 1,
  mean.trt,
  sd.trt = 1,
  mean.bg = 0,
  sd.bg = 0.5,
  N,
  p_total,
  p_trt_relevant,
  p_bat_relevant,
  percentage_overlap_samples = 0.5,
  percentage_overlap_variables = 0.5,
  data.cor,
  disp = 10,
  prob_zero = 0,
  seeds = round(runif(1, 0, 999))
)

Arguments

batch.group

Integer, the number of batch groups. Only valid for 2 or 3 batch groups. Default value is 2.

mean.batch

Numeric, the mean of batch effect coefficients for all related variables.

sd.batch

Numeric, the variability (standard deviation) of batch effect coefficients for all related variables.

mean.trt

Numeric, the mean of treatment effect coefficients for all related variables

sd.trt

Numeric, the variability (standard deviation) of treatment effect coefficients for all related variables.

mean.bg

Numeric, the mean of background noise.

sd.bg

Numeric, the variability (standard deviation) of background noise.

N

Integer, the number of samples.

p_total

Integer, the number of variables.

p_trt_relevant

Integer, the number of variables affected by treatment effects.

p_bat_relevant

Integer, the number of variables affected by batch effects.

percentage_overlap_samples

Numeric, in the range of \(0\) to \(1\), the percentage of samples from one batch group in one treatment group.

percentage_overlap_variables

Numeric, in the range of \(0\) to \(1\), the percentage of variables with both batch and treatment effects given fixed number of variables with each effect.

data.cor

Correlation matrix. The correlation between each pair of variables.

disp

Numeric. The dispersion parameter of negative binomial distribution.

prob_zero

Numeric, in the range of \(0\) to \(1\). The probability of each value to be 0.

seeds

Integer. Default value is randomly selected from \(0\) to \(999\).

Value

simData_mnegbinom returns a list that contains the following components:

data

The simulated data.

cleanData

The ground-truth data without the batch effect.

data_mu_matrix

The \(ln\) value of the mean to model negative binomial distribution for the simulated data.

clean_mu_matrix

The \(ln\) value of the mean to model negative binomial distribution for the ground-truth data.

Y.trt

The simulated outcome variable indicating the treatment grouping information.

Y.bat

The simulated outcome variable indicating the batch grouping information.

true.trt

The variables assigned with the treatment effect.

true.batch

The variables assigned with the batch effect.

zero_index

The indexes of cells containing the resulting counts from the Multivariate Negative Binomial Distribution that are inflated as zero.

References

Wang Y, Lê Cao K (2020). “A multivariate method to correct for batch effects in microbiome data.” bioRxiv.

Author

Yiwen Wang, Kim-Anh Lê Cao

Examples

## Not run:
## First example
## To simulate the correlation structure
data.cor.res = corStruct(p = 300, zero_prob = 0.7)

## Balanced batch x treatment design
simulation <- simData_mnegbinom(mean.batch = 7,
                                sd.batch = 8,
                                mean.trt = 3,
                                sd.trt = 2,
                                mean.bg = 0,
                                sd.bg = 0.2,
                                N = 40,
                                p_total = 300,
                                p_trt_relevant = 100,
                                p_bat_relevant = 200,
                                percentage_overlap_samples = 0.5,
                                percentage_overlap_variables = 0.5,
                                data.cor = data.cor.res$data.cor,
                                disp = 10,
                                prob_zero = 0,
                                seeds = 1)

## Second example
## Unbalanced batch x treatment design
simulation <- simData_mnegbinom(mean.batch = 7,
                                sd.batch = 8,
                                mean.trt = 3,
                                sd.trt = 2,
                                mean.bg = 0,
                                sd.bg = 0.2,
                                N = 40,
                                p_total = 300,
                                p_trt_relevant = 100,
                                p_bat_relevant = 200,
                                percentage_overlap_samples = 0.2,
                                percentage_overlap_variables = 0.5,
                                data.cor = data.cor.res$data.cor,
                                disp = 10,
                                prob_zero = 0,
                                seeds = 1)
## End(Not run)